// src/api/models/subscriptionSnapshotListItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Partners
 * Allta Partners 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */
import type { SubscriptionSubscriptionItem } from './subscriptionSubscriptionItem';

export interface SubscriptionSnapshotListItem {
  /** ID */
  id: string;
  /** 구독권 종류 */
  type: string;
  /**
   * 구독권 스냅샷 상태
   * @nullable
   */
  status?: string | null;
  /** 서비스 종류 */
  serviceType: string;
  /**
   * 최대 사용 횟수
   * @nullable
   */
  maxUsage?: number | null;
  /** 사용 횟수 */
  usage: number;
  /** 차량 번호 */
  carNumber: string;
  /** 생성일 */
  createdAt: string;
  /** 구독권 정보 */
  subscription: SubscriptionSubscriptionItem;
}